+++ imod/formats/prj/disv_conversion.py
@@ -20,11 +20,12 @@ from imod.mf6.model import Modflow6Model
 from imod.mf6.utilities.package_utils import get_repeat_stress
 from imod.prepare.layer import get_upper_active_grid_cells
 from imod.typing import GridDataArray
+from imod.util.imports import MissingOptionalModule
 
 try:
     import geopandas as gpd
 except ImportError:
-    gpd = imod.util.MissingOptionalModule("geopandas")
+    gpd = MissingOptionalModule("geopandas")
 
 
 def hash_xy(da: xr.DataArray) -> Tuple[int]: