+++ imod/formats/rasterio.py
@@ -13,15 +13,16 @@ from typing import Dict
 import numpy as np
 import pandas as pd
 
-from imod import util
+import imod
 from imod.formats import array_io
+from imod.util.imports import MissingOptionalModule
 
 # since rasterio is a big dependency that is sometimes hard to install
 # and not always required, we made this an optional dependency
 try:
     import rasterio
 except ImportError:
-    rasterio = util.MissingOptionalModule("rasterio")
+    rasterio = MissingOptionalModule("rasterio")
 
 f_open = open
 
@@ -130,7 +131,7 @@ def _limitations(riods, path):
 
 
 def header(path, pattern):
-    attrs = util.decompose(path, pattern)
+    attrs = imod.util.path.decompose(path, pattern)
 
     # TODO:
     # Check bands, rotation, etc.
@@ -431,7 +432,7 @@ def write(path, da, driver=None, nodata=np.nan, dtype=None):
     if extradims:
         raise ValueError(f"Only x and y dimensions supported, found {da.dims}")
     # transform will be affine object in next xarray
-    profile["transform"] = util.transform(da)
+    profile["transform"] = imod.util.spatial.transform(da)
     profile["driver"] = driver
     profile["height"] = da.y.size
     profile["width"] = da.x.size
@@ -441,7 +442,7 @@ def write(path, da, driver=None, nodata=np.nan, dtype=None):
 
     # Allow writing ASCII grids even if rasterio isn't installed. This is
     # useful for e.g. MetaSWAP input.
-    if isinstance(rasterio, util.MissingOptionalModule) and driver == "AAIGrid":
+    if isinstance(rasterio, MissingOptionalModule) and driver == "AAIGrid":
         write_aaigrid(path, da.values, profile)
     else:
         with rasterio.Env():
@@ -458,7 +459,7 @@ def save(path, a, driver=None, nodata=np.nan, pattern=None, dtype=None):
     written, like the ``imod.rasterio.write`` function. This function is more general
     and also supports ``time`` and ``layer`` and other dimensions. It will split these up,
     give them their own filename according to the conventions in
-    ``imod.util.compose``, and write them each.
+    ``imod.util.path.compose``, and write them each.
 
     Parameters
     ----------