+++ imod/mf6/pkgbase.py
@@ -112,7 +112,7 @@ class PackageBase(IPackageBase, abc.ABC):
 
         if dataset.ugrid_roles.topology:
             dataset = xu.UgridDataset(dataset)
-            dataset = imod.util.from_mdal_compliant_ugrid2d(dataset)
+            dataset = imod.util.spatial.from_mdal_compliant_ugrid2d(dataset)
 
         # Replace NaNs by None
         for key, value in dataset.items():