+++ imod/tests/test_mf6/test_mf6_transport_model.py
@@ -20,7 +20,7 @@ def test_long_package_name():
 
 def test_transport_model_rendering():
     adv = imod.mf6.AdvectionCentral()
-    disp = imod.mf6.Dispersion(1e-4, 1.0, 10.0, 1.0, 2.0, 3.0, True, True)
+    disp = imod.mf6.Dispersion(1e-4, 1.0, 10.0, 1.0, 2.0, 3.0, False, True)
     m = imod.mf6.GroundwaterTransportModel(print_input=True, save_flows=True)
     m["dsp"] = disp
     m["adv"] = adv