+++ imod/tests/test_mf6/test_multimodel/test_exchange_creator_structured.py
@@ -8,7 +8,7 @@ from imod.mf6.multimodel.exchange_creator_structured import ExchangeCreator_Stru
 from imod.mf6.multimodel.modelsplitter import create_partition_info
 from imod.tests.fixtures.flow_basic_fixture import BasicDisSettings
 from imod.typing.grid import zeros_like
-from imod.util import spatial_reference
+from imod.util.spatial import spatial_reference
 
 ExpectedExchanges = namedtuple("ExpectedExchanges", "cell_id1 cell_id2 cl1 cl2 hwva")