+++ imod/tests/test_util.py
@@ -11,7 +11,7 @@ import pytest
 import xarray as xr
 import xugrid as xu
 
-from imod import util
+import imod
 
 
 @pytest.fixture(scope="function")
@@ -60,28 +60,28 @@ def test_compose():
         "time": datetime.datetime(2018, 2, 22, 9, 6, 57),
         "species": 6,
     }
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_c6_20180222090657_l5.idf")
     assert path == targetpath
 
     d.pop("species")
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_20180222090657_l5.idf")
     assert path == targetpath
 
     d.pop("layer")
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_20180222090657.idf")
     assert path == targetpath
 
     d.pop("time")
     d["layer"] = 1
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_l1.idf")
     assert path == targetpath
 
     d["species"] = 6
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_c6_l1.idf")
     assert path == targetpath
 
@@ -96,26 +96,26 @@ def test_compose__pattern():
     targetpath = pathlib.Path(d["directory"], "head_2018-02-22_l05.foo")
 
     d["time"] = datetime.datetime(2018, 2, 22, 9, 6, 57)
-    path = util.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
+    path = imod.util.path.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
     assert path == targetpath
 
     d["time"] = cftime.DatetimeProlepticGregorian(2018, 2, 22, 9, 6, 57)
-    path = util.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
+    path = imod.util.path.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
     assert path == targetpath
 
     d["time"] = np.datetime64("2018-02-22 09:06:57")
-    path = util.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
+    path = imod.util.path.compose(d, pattern="{name}_{time:%Y-%m-%d}_l{layer:02d}{extension}")
     assert path == targetpath
 
     targetpath = pathlib.Path(d["directory"], ".foo_makes_head_no_layer5_sense_day22")
-    path = util.compose(
+    path = imod.util.path.compose(
         d, pattern="{extension}_makes_{name}_no_layer{layer:d}_sense_day{time:%d}"
     )
     assert path == targetpath
 
 
 def test_decompose():
-    d = util.decompose("path/to/head_20180222090657_l5.idf")
+    d = imod.util.path.decompose("path/to/head_20180222090657_l5.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -129,7 +129,7 @@ def test_decompose():
 
 
 def test_decompose_species():
-    d = util.decompose("path/to/conc_c3_20180222090657_l5.idf")
+    d = imod.util.path.decompose("path/to/conc_c3_20180222090657_l5.idf")
     refd = {
         "extension": ".idf",
         "species": 3,
@@ -144,7 +144,7 @@ def test_decompose_species():
 
 
 def test_decompose_short_date():
-    d = util.decompose("path/to/head_20180222_l5.idf")
+    d = imod.util.path.decompose("path/to/head_20180222_l5.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -158,7 +158,7 @@ def test_decompose_short_date():
 
 
 def test_decompose_nonstandard_date():
-    d = util.decompose("path/to/head_2018-02-22_l5.idf")
+    d = imod.util.path.decompose("path/to/head_2018-02-22_l5.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -172,7 +172,7 @@ def test_decompose_nonstandard_date():
 
 
 def test_decompose_only_year():
-    d = util.decompose("path/to/head_2018_l5.idf", pattern="{name}_{time}_l{layer}")
+    d = imod.util.path.decompose("path/to/head_2018_l5.idf", pattern="{name}_{time}_l{layer}")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -186,7 +186,7 @@ def test_decompose_only_year():
 
 
 def test_decompose_underscore():
-    d = util.decompose("path/to/starting_head_20180222090657_l5.idf")
+    d = imod.util.path.decompose("path/to/starting_head_20180222090657_l5.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -200,7 +200,7 @@ def test_decompose_underscore():
 
 
 def test_decompose_dash():
-    d = util.decompose("path/to/starting-head_20180222090657_l5.idf")
+    d = imod.util.path.decompose("path/to/starting-head_20180222090657_l5.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -214,7 +214,7 @@ def test_decompose_dash():
 
 
 def test_decompose_steady_state():
-    d = util.decompose("path/to/head_steady-state_l64.idf")
+    d = imod.util.path.decompose("path/to/head_steady-state_l64.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -228,7 +228,7 @@ def test_decompose_steady_state():
 
 
 def test_decompose_underscore_in_name():
-    d = util.decompose("path/to/some_name.idf")
+    d = imod.util.path.decompose("path/to/some_name.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("path", "to"),
@@ -240,7 +240,7 @@ def test_decompose_underscore_in_name():
 
 
 def test_decompose_pattern_underscore():
-    d = util.decompose(
+    d = imod.util.path.decompose(
         "path/to/starting_head_20180222090657_l5.idf", pattern="{name}_{time}_l{layer}"
     )
     refd = {
@@ -256,7 +256,7 @@ def test_decompose_pattern_underscore():
 
 
 def test_decompose_pattern_dash():
-    d = util.decompose(
+    d = imod.util.path.decompose(
         "path/to/starting-head_20180222090657_l5.idf", pattern="{name}_{time}_l{layer}"
     )
     refd = {
@@ -273,7 +273,7 @@ def test_decompose_pattern_dash():
 
 def test_decompose_regexpattern():
     pattern = re.compile(r"(?P<name>[\w]+)L(?P<layer>[\d+]*)", re.IGNORECASE)
-    d = util.decompose("headL11.idf", pattern=pattern)
+    d = imod.util.path.decompose("headL11.idf", pattern=pattern)
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -286,7 +286,7 @@ def test_decompose_regexpattern():
 
 
 def test_decompose_nodate():
-    d = util.decompose("dem_10m.idf")
+    d = imod.util.path.decompose("dem_10m.idf")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -298,7 +298,7 @@ def test_decompose_nodate():
 
 
 def test_decompose_dateonly():
-    d = util.decompose("20180222090657.idf", pattern="{time}")
+    d = imod.util.path.decompose("20180222090657.idf", pattern="{time}")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -311,7 +311,7 @@ def test_decompose_dateonly():
 
 
 def test_decompose_datelayeronly():
-    d = util.decompose("20180222090657_l7.idf", pattern="{time}_l{layer}")
+    d = imod.util.path.decompose("20180222090657_l7.idf", pattern="{time}_l{layer}")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -325,7 +325,7 @@ def test_decompose_datelayeronly():
 
 
 def test_decompose_z_float():
-    d = util.decompose("test_0.25.idf", pattern="{name}_{z}")
+    d = imod.util.path.decompose("test_0.25.idf", pattern="{name}_{z}")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -346,13 +346,13 @@ def test_compose_year9999():
         "time": datetime.datetime(9999, 2, 22, 9, 6, 57),
         "dims": ["time"],
     }
-    path = util.compose(d)
+    path = imod.util.path.compose(d)
     targetpath = pathlib.Path(d["directory"], "head_99990222090657_l5.idf")
     assert path == targetpath
 
 
 def test_decompose_dateonly_year9999():
-    d = util.decompose("99990222090657.idf", pattern="{time}")
+    d = imod.util.path.decompose("99990222090657.idf", pattern="{time}")
     refd = {
         "extension": ".idf",
         "directory": pathlib.Path("."),
@@ -366,7 +366,7 @@ def test_decompose_dateonly_year9999():
 
 def test_datetime_conversion__withinbounds():
     times = [datetime.datetime(y, 1, 1) for y in range(2000, 2010)]
-    converted, use_cftime = util._convert_datetimes(times, use_cftime=False)
+    converted, use_cftime = imod.util.time._convert_datetimes(times, use_cftime=False)
     assert use_cftime is False
     assert all(t.dtype == "<M8[ns]" for t in converted)
     assert converted[0] == np.datetime64("2000-01-01", "ns")
@@ -376,7 +376,7 @@ def test_datetime_conversion__withinbounds():
 def test_datetime_conversion__outofbounds():
     times = [datetime.datetime(y, 1, 1) for y in range(1670, 1680)]
     with pytest.warns(UserWarning):
-        converted, use_cftime = util._convert_datetimes(times, use_cftime=False)
+        converted, use_cftime = imod.util.time._convert_datetimes(times, use_cftime=False)
     assert use_cftime is True
     assert all(type(t) is cftime.DatetimeProlepticGregorian for t in converted)
     assert converted[0] == cftime.DatetimeProlepticGregorian(1670, 1, 1)
@@ -385,7 +385,7 @@ def test_datetime_conversion__outofbounds():
 
 def test_datetime_conversion__withinbounds_cftime():
     times = [datetime.datetime(y, 1, 1) for y in range(2000, 2010)]
-    converted, use_cftime = util._convert_datetimes(times, use_cftime=True)
+    converted, use_cftime = imod.util.time._convert_datetimes(times, use_cftime=True)
     assert use_cftime is True
     assert all(type(t) is cftime.DatetimeProlepticGregorian for t in converted)
     assert converted[0] == cftime.DatetimeProlepticGregorian(2000, 1, 1)
@@ -398,7 +398,7 @@ def test_transform():
     coords = {"x": [0.5, 1.5, 2.5], "y": [1.5, 0.5]}
     dims = ("y", "x")
     da = xr.DataArray(data, coords, dims)
-    actual = util.transform(da)
+    actual = imod.util.spatial.transform(da)
     expected = affine.Affine(1.0, 0.0, 0.0, 0.0, -1.0, 2.0)
     assert actual == expected
 
@@ -411,7 +411,7 @@ def test_transform():
     }
     dims = ("y", "x")
     da = xr.DataArray(data, coords, dims)
-    actual = util.transform(da)
+    actual = imod.util.spatial.transform(da)
     assert actual == expected
 
     # explicit dx dy, non-equidistant
@@ -424,54 +424,54 @@ def test_transform():
     dims = ("y", "x")
     da = xr.DataArray(data, coords, dims)
     with pytest.raises(ValueError):
-        util.transform(da)
+        imod.util.spatial.transform(da)
 
 
 def test_is_divisor():
     a = np.array([1.0, 0.5, 0.1])
     b = 0.05
-    assert util.is_divisor(a, b)
-    assert util.is_divisor(-a, b)
-    assert util.is_divisor(a, -b)
-    assert util.is_divisor(-a, -b)
+    assert imod.util.spatial.is_divisor(a, b)
+    assert imod.util.spatial.is_divisor(-a, b)
+    assert imod.util.spatial.is_divisor(a, -b)
+    assert imod.util.spatial.is_divisor(-a, -b)
     b = 0.07
-    assert not util.is_divisor(a, b)
-    assert not util.is_divisor(-a, b)
-    assert not util.is_divisor(a, -b)
-    assert not util.is_divisor(-a, -b)
+    assert not imod.util.spatial.is_divisor(a, b)
+    assert not imod.util.spatial.is_divisor(-a, b)
+    assert not imod.util.spatial.is_divisor(a, -b)
+    assert not imod.util.spatial.is_divisor(-a, -b)
 
 
 def test_empty():
-    da = util.empty_2d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0)
+    da = imod.util.spatial.empty_2d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0)
     assert da.isnull().all()
     assert np.allclose(da["x"], [0.5, 1.5])
     assert np.allclose(da["y"], [11.5, 10.5])
     assert da.dims == ("y", "x")
     # Sign on dx, dy should be ignored
-    da = util.empty_2d(-1.0, 0.0, 2.0, 1.0, 10.0, 12.0)
+    da = imod.util.spatial.empty_2d(-1.0, 0.0, 2.0, 1.0, 10.0, 12.0)
     assert np.allclose(da["x"], [0.5, 1.5])
     assert np.allclose(da["y"], [11.5, 10.5])
 
     with pytest.raises(ValueError, match="layer must be 1d"):
-        util.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [[1, 2]])
+        imod.util.spatial.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [[1, 2]])
 
-    da3d = util.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, 1)
+    da3d = imod.util.spatial.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, 1)
     assert da3d.ndim == 3
-    da3d = util.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [1, 3])
+    da3d = imod.util.spatial.empty_3d(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [1, 3])
     assert np.array_equal(da3d["layer"], [1, 3])
     assert da3d.dims == ("layer", "y", "x")
 
     times = ["2000-01-01", "2001-01-01"]
     with pytest.raises(ValueError, match="time must be 1d"):
-        util.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [times])
+        imod.util.spatial.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [times])
 
-    da2dt = util.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, times[0])
+    da2dt = imod.util.spatial.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, times[0])
     assert da2dt.ndim == 3
-    da2dt = util.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, times)
+    da2dt = imod.util.spatial.empty_2d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, times)
     assert isinstance(da2dt["time"].values[0], np.datetime64)
     assert da2dt.dims == ("time", "y", "x")
 
-    da3dt = util.empty_3d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [0, 1], times)
+    da3dt = imod.util.spatial.empty_3d_transient(1.0, 0.0, 2.0, -1.0, 10.0, 12.0, [0, 1], times)
     assert da3dt.ndim == 4
     assert da3dt.dims == ("time", "layer", "y", "x")
 
@@ -483,21 +483,21 @@ def test_where():
         ["y", "x"],
     )
     cond = a <= 1
-    actual = util.where(cond, if_true=a, if_false=1.0)
+    actual = imod.util.structured.where(cond, if_true=a, if_false=1.0)
     assert np.allclose(actual.values, [[0.0, 1.0], [1.0, np.nan]], equal_nan=True)
 
-    actual = util.where(cond, if_true=0.0, if_false=1.0)
+    actual = imod.util.structured.where(cond, if_true=0.0, if_false=1.0)
     assert np.allclose(actual.values, [[0.0, 0.0], [1.0, 1.0]], equal_nan=True)
 
-    actual = util.where(cond, if_true=a, if_false=1.0, keep_nan=False)
+    actual = imod.util.structured.where(cond, if_true=a, if_false=1.0, keep_nan=False)
     assert np.allclose(actual.values, [[0.0, 1.0], [1.0, 1.0]])
 
     with pytest.raises(ValueError, match="at least one of"):
-        util.where(False, 1, 0)
+        imod.util.structured.where(False, 1, 0)
 
 
 def test_compliant_ugrid2d(ugrid_ds, write=False):
-    uds = util.mdal_compliant_ugrid2d(ugrid_ds)
+    uds = imod.util.spatial.mdal_compliant_ugrid2d(ugrid_ds)
 
     assert isinstance(uds, xr.Dataset)
     for i in range(1, 4):
@@ -520,10 +520,10 @@ def test_compliant_ugrid2d(ugrid_ds, write=False):
 
 
 def test_mdal_compliant_roundtrip(ugrid_ds):
-    uds = xu.UgridDataset(util.mdal_compliant_ugrid2d(ugrid_ds))
+    uds = xu.UgridDataset(imod.util.spatial.mdal_compliant_ugrid2d(ugrid_ds))
     uds["b"] = (("time", "layer"), np.random.rand(5, 3))
     uds["c"] = (("layer", "mesh2d_nFaces"), np.random.rand(3, 4))
-    back = util.from_mdal_compliant_ugrid2d(uds)
+    back = imod.util.spatial.from_mdal_compliant_ugrid2d(uds)
 
     assert isinstance(back, xu.UgridDataset)
     assert back["a"].dims == ("time", "layer", "mesh2d_nFaces")
@@ -533,7 +533,7 @@ def test_mdal_compliant_roundtrip(ugrid_ds):
 
 
 def test_to_ugrid2d(write=False):
-    a2d = util.empty_2d(
+    a2d = imod.util.spatial.empty_2d(
         dx=1.0,
         xmin=0.0,
         xmax=2.0,
@@ -543,15 +543,15 @@ def test_to_ugrid2d(write=False):
     )
 
     with pytest.raises(TypeError, match="data must be xarray"):
-        util.to_ugrid2d(a2d.values)
+        imod.util.spatial.to_ugrid2d(a2d.values)
     with pytest.raises(ValueError, match="A name is required"):
-        util.to_ugrid2d(a2d)
+        imod.util.spatial.to_ugrid2d(a2d)
 
     a2d.name = "a"
     with pytest.raises(ValueError, match="Last two dimensions of da"):
-        util.to_ugrid2d(a2d.transpose())
+        imod.util.spatial.to_ugrid2d(a2d.transpose())
 
-    uds = util.to_ugrid2d(a2d)
+    uds = imod.util.spatial.to_ugrid2d(a2d)
     assert isinstance(uds, xr.Dataset)
     assert "a" in uds
 
@@ -572,7 +572,7 @@ def test_to_ugrid2d(write=False):
     ds = xr.Dataset()
     ds["a"] = a2d
     ds["b"] = a2d.copy()
-    uds = util.to_ugrid2d(ds)
+    uds = imod.util.spatial.to_ugrid2d(ds)
     assert "a" in uds
     assert "b" in uds
     assert isinstance(uds, xr.Dataset)
@@ -581,7 +581,7 @@ def test_to_ugrid2d(write=False):
         uds.to_netcdf("ugrid-a2d-ds.nc")
 
     # transient 2d
-    a2dt = util.empty_2d_transient(
+    a2dt = imod.util.spatial.empty_2d_transient(
         dx=1.0,
         xmin=0.0,
         xmax=2.0,
@@ -591,7 +591,7 @@ def test_to_ugrid2d(write=False):
         time=pd.date_range("2000-01-01", "2000-01-05"),
     )
     a2dt.name = "a"
-    uds = util.to_ugrid2d(a2dt)
+    uds = imod.util.spatial.to_ugrid2d(a2dt)
     assert "a" in uds
     assert uds["time"].encoding["dtype"] == np.float64
 
@@ -599,7 +599,7 @@ def test_to_ugrid2d(write=False):
         uds.to_netcdf("ugrid-a2dt.nc")
 
     # 3d
-    a3d = util.empty_3d(
+    a3d = imod.util.spatial.empty_3d(
         dx=1.0,
         xmin=0.0,
         xmax=2.0,
@@ -609,7 +609,7 @@ def test_to_ugrid2d(write=False):
         layer=[1, 2, 3],
     )
     a3d.name = "a"
-    uds = util.to_ugrid2d(a3d)
+    uds = imod.util.spatial.to_ugrid2d(a3d)
     assert isinstance(uds, xr.Dataset)
     for i in range(1, 4):
         assert f"a_layer_{i}" in uds
@@ -618,7 +618,7 @@ def test_to_ugrid2d(write=False):
         uds.to_netcdf("ugrid-a3d.nc")
 
     # transient 3d
-    a3dt = util.empty_3d_transient(
+    a3dt = imod.util.spatial.empty_3d_transient(
         dx=1.0,
         xmin=0.0,
         xmax=2.0,
@@ -629,7 +629,7 @@ def test_to_ugrid2d(write=False):
         time=pd.date_range("2000-01-01", "2000-01-05"),
     )
     a3dt.name = "a"
-    uds = util.to_ugrid2d(a3dt)
+    uds = imod.util.spatial.to_ugrid2d(a3dt)
     for i in range(1, 4):
         assert f"a_layer_{i}" in uds
     assert uds["time"].encoding["dtype"] == np.float64
@@ -641,57 +641,84 @@ def test_to_ugrid2d(write=False):
 def test_replace():
     # replace scalar
     da = xr.DataArray([0, 1, 2])
-    out = util.replace(da, 1, 10)
+    out = imod.util.structured.replace(da, 1, 10)
     assert out.equals(xr.DataArray([0, 10, 2]))
 
     # Replace NaN by scalar
     da = xr.DataArray([np.nan, 1.0, 2.0])
-    out = util.replace(da, np.nan, 10.0)
+    out = imod.util.structured.replace(da, np.nan, 10.0)
     assert out.equals(xr.DataArray([10.0, 1.0, 2.0]))
 
     # replace two
     da = xr.DataArray([0, 1, 2])
-    out = util.replace(da, [1, 2], [10, 20])
+    out = imod.util.structured.replace(da, [1, 2], [10, 20])
     assert out.equals(xr.DataArray([0, 10, 20]))
 
     # With a NaN in the data
     da = xr.DataArray([np.nan, 1.0, 2.0])
-    out = util.replace(da, [1, 2], [10, 20])
+    out = imod.util.structured.replace(da, [1, 2], [10, 20])
     assert out.equals(xr.DataArray([np.nan, 10.0, 20.0]))
 
     # Replace a NaN value
     da = xr.DataArray([np.nan, 1.0, 2.0])
-    out = util.replace(da, [np.nan, 2], [10, 20])
+    out = imod.util.structured.replace(da, [np.nan, 2], [10, 20])
     assert out.equals(xr.DataArray([10.0, 1.0, 20.0]))
 
     # With non-present values in to_replace
     da = xr.DataArray([np.nan, 1.0, 1.0, 2.0])
-    out = util.replace(da, [1.0, 2.0, 30.0], [10.0, 20.0, 30.0])
+    out = imod.util.structured.replace(da, [1.0, 2.0, 30.0], [10.0, 20.0, 30.0])
     assert out.equals(xr.DataArray([np.nan, 10.0, 10.0, 20.0]))
 
     # With a nan and non-present values
     da = xr.DataArray([np.nan, 1.0, 1.0, 2.0])
-    out = util.replace(da, [np.nan, 1.0, 2.0, 30.0], 10.0)
+    out = imod.util.structured.replace(da, [np.nan, 1.0, 2.0, 30.0], 10.0)
     assert out.equals(xr.DataArray([10.0, 10.0, 10.0, 10.0]))
 
     # With a dask array
     da = xr.DataArray(dask.array.full(3, 1.0))
-    out = util.replace(da, [1.0, 2.0], [10.0, 20.0])
+    out = imod.util.structured.replace(da, [1.0, 2.0], [10.0, 20.0])
     assert isinstance(out.data, dask.array.Array)
     assert out.equals(xr.DataArray([10.0, 10.0, 10.0]))
 
     # scalar to_replace, non-scalar value
     with pytest.raises(TypeError):
-        util.replace(da, 1.0, [10.0, 20.0])
+        imod.util.structured.replace(da, 1.0, [10.0, 20.0])
 
     # 2D arrays
     with pytest.raises(ValueError):
-        util.replace(da, [[1.0, 2.0]], [[10.0, 20.0]])
+        imod.util.structured.replace(da, [[1.0, 2.0]], [[10.0, 20.0]])
 
     # 1D to_replace, 2D value
     with pytest.raises(ValueError):
-        util.replace(da, [1.0, 2.0], [[10.0, 20.0]])
+        imod.util.structured.replace(da, [1.0, 2.0], [[10.0, 20.0]])
 
     # 1D, different size
     with pytest.raises(ValueError):
-        util.replace(da, [1.0, 2.0], [10.0, 20.0, 30.0])
+        imod.util.structured.replace(da, [1.0, 2.0], [10.0, 20.0, 30.0])
+
+
+def test_public_api():
+    """
+    Test if functions previously in imod.util.py are still available under same
+    namespace
+    """
+    from imod.util import (
+        cd,  # noqa: F401
+        empty_2d,  # noqa: F401
+        empty_2d_transient,  # noqa: F401
+        empty_3d,  # noqa: F401
+        empty_3d_transient,  # noqa: F401
+        from_mdal_compliant_ugrid2d,  # noqa: F401
+        ignore_warnings,  # noqa: F401
+        mdal_compliant_ugrid2d,  # noqa: F401
+        replace,  # noqa: F401
+        round_extent,  # noqa: F401
+        spatial_reference,  # noqa: F401
+        temporary_directory,  # noqa: F401
+        to_datetime,  # noqa: F401
+        to_ugrid2d,  # noqa: F401
+        transform,  # noqa: F401
+        ugrid2d_data,  # noqa: F401
+        values_within_range,  # noqa: F401
+        where,  # noqa: F401
+    )